Allow keeping logs "forever"
Where we define "forever" as a number of days large enough to definitely
outlive the device the app runs on (100 000 days = ~274 years).

Size should not be a problem: even at an unlikely high device lifetime
of 10 years, the database would grow to a size of only ~5.7MiB based on
my observations.